refactor(cli): reuse UploadProgressTracker for file uploads
Move UploadProgressTracker/DownloadProgressTracker into shared CLI
components and wire files/fine-tuning upload+download through them.
Also validate upload redirect URLs/file IDs to mitigate SSRF.
